projects/memory-chat-persistent/pyproject.toml
[project]
name = "memory-chat"
version = "0.1.0"
description = "Minimal LLM chat app with provider/model selection and configurable conversation memory"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.35.0",
    "openai>=1.30.0",
    "anthropic>=0.28.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.73.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
